The gumball and friends hold still far from the origin too
The mesh fix gave far geometry an anchor; everything drawn over it
still went to the GPU as absolute float32 and trembled against the now
solid model. Overlays re-upload every frame, so unlike a mesh they can
all share one moving anchor: where the camera is looking, which is
where everything worth overlaying is.

_paint_frame folds view_anchor(camera.target) into the one float32
matrix every overlay receives, and each overlay rebases what it
uploads to match: the rubber band and its markers, snap markers, the
axis guide, the control-point display, curvature combs, direction
arrows, ghosts, image planes, and the gumball through its _lines and
_tris funnel. On paper the anchor is None and rebased is a plain cast.

The grid is its own case: its geometry is cplane-local, so the
cplane's basis is that pass's anchor, folded in float64 before the
cast. CPlane.basis_matrix was a float32 container, rounding a far
plane's origin at birth the way utils/math3d.py used to, and is now
float64.

Near the origin the anchor is None everywhere and the frame is
bit-for-bit what it was, which the probe on Xephyr confirmed: mean
pixel churn under a slow orbit identical to four decimal places, while
at 500,000 units the churn dropped from 0.106 to 0.092, the remainder
being the orbit itself.

+++ serpentine3d/core/cplane.py
@@ -45,7 +45,10 @@ class CPlane:
 
     def basis_matrix(self) -> np.ndarray:
         """4x4 model matrix mapping plane-local coords to world."""
-        m = np.eye(4, dtype=np.float32)
+        # float64: a plane out at survey coordinates would otherwise have
+        # its origin rounded to float32's 3cm grid at birth, before the
+        # viewport's own float64 fold could do anything about it.
+        m = np.eye(4)
         m[:3, 0] = self.xdir
         m[:3, 1] = self.ydir
         m[:3, 2] = self.normal
